lint-and-fix prettier shelljs from github

+++ src/commands/remove.ts
@@ -8,7 +8,7 @@ import {Command, flags} from '@oclif/command'
 import Utils from '../classes/utils'
 import Pacman from '../classes/pacman'
 import {IInstall} from '../interfaces'
-import {execSync} from 'child_process'
+import {execSync} from 'node:child_process'
 
 import {exec} from '../lib/utils'
 
@@ -18,11 +18,7 @@ import {exec} from '../lib/utils'
 export default class Remove extends Command {
    static description = 'remove eggs and others stuff'
 
-   static examples = [
-     '$ sudo eggs remove \nremove eggs\n',
-     '$ sudo eggs remove --purge \nremove eggs, eggs configurations, configuration\'s files\n',
-     '$ sudo eggs remove --autoremove \nremove eggs, eggs configurations, packages dependencies\n',
-   ]
+   static examples = ['$ sudo eggs remove \nremove eggs\n', "$ sudo eggs remove --purge \nremove eggs, eggs configurations, configuration's files\n", '$ sudo eggs remove --autoremove \nremove eggs, eggs configurations, packages dependencies\n']
 
    static flags = {
      purge: flags.boolean({char: 'p', description: 'remove eggs configurations files'}),
@@ -42,59 +38,54 @@ export default class Remove extends Command {
 
      if (Utils.isRoot()) {
        /**
-         * debian package
-         */
-       if (Utils.isDebPackage()) {
-         if (await Utils.customConfirm()) {
+          * debian package
+          */
+       if (Utils.isDebPackage() && await Utils.customConfirm()) {
+         /**
+                * in caso di autoremove, tolgo le dipendenze di versione PRIMA di eggs, altrimenti si inchioda
+                */
+         if (flags.autoremove) {
            /**
-               * in caso di autoremove, tolgo le dipendenze di versione PRIMA di eggs, altrimenti si inchioda
-               */
-           if (flags.autoremove) {
-             /**
                    * Rimozione dipendenze da versione live-config / open-infrastructure-system-config (ubuntu bionic)
                    */
-             if (Pacman.packageIsInstalled('open-infrastructure-system-config')) {
-               execSync('apt-get purge open-infrastructure-system-config --yes')
-             } else if (Pacman.packageIsInstalled('live-config')) {
-               execSync('apt-get purge live-config --yes')
-             }
+           if (Pacman.packageIsInstalled('open-infrastructure-system-config')) {
+             execSync('apt-get purge open-infrastructure-system-config --yes')
+           } else if (Pacman.packageIsInstalled('live-config')) {
+             execSync('apt-get purge live-config --yes')
            }
+         }
 
-           if (flags.purge) {
-             execSync('apt-get purge eggs --yes')
-           } else {
-             execSync('apt-get remove eggs --yes')
-           }
+         if (flags.purge) {
+           execSync('apt-get purge eggs --yes')
+         } else {
+           execSync('apt-get remove eggs --yes')
+         }
 
-           if (flags.autoremove) {
-             execSync('apt-get autoremove --yes')
-           }
+         if (flags.autoremove) {
+           execSync('apt-get autoremove --yes')
          }
        }
 
        /**
-         * sources
-         */
-       if (Utils.isSources()) {
-         if (await Utils.customConfirm()) {
-           if (flags.autoremove) {
-             // await Pacman.prerequisitesRemove()
-           }
-
-           if (flags.purge) {
-             await Pacman.configurationRemove()
-
-             // eggs completion
-             execSync('rm -f /etc/bash_completion.d/eggs.bash')
-
-             // manpages
-             execSync('rm -f /usr/share/man/man1/eggs.1.gz')
-           }
-
-           Utils.warning('You are using  eggs as sources. I\'ll NOT remove it')
+          * sources
+          */
+       if (Utils.isSources() && await Utils.customConfirm()) {
+         if (flags.autoremove) {
+           // await Pacman.prerequisitesRemove()
          }
+
+         if (flags.purge) {
+           await Pacman.configurationRemove()
+
+           // eggs completion
+           execSync('rm -f /etc/bash_completion.d/eggs.bash')
+
+           // manpages
+           execSync('rm -f /usr/share/man/man1/eggs.1.gz')
+         }
+
+         Utils.warning("You are using  eggs as sources. I'll NOT remove it")
        }
      }
    }
 }
-
